2005-01-06  Sureshkumar T  <tsureshkumar@novell.com>
	* Tds.cs: change the typecast to ushort for
	GetDateTimeValue. fixes bug #70651.


svn path=/trunk/mcs/; revision=38422

+++ mcs/class/Mono.Data.Tds/Mono.Data.Tds.Protocol/Tds.cs
@@ -572,7 +572,9 @@ namespace Mono.Data.Tds.Protocol {
 				}
 				break;
 			case 4 :
-				result = epoch.AddDays ((int) comm.GetTdsShort ());
+                                // MSDN says small datetime is stored in 2 bytes as no of days
+                                // *after* 1/1/1900. so, cast to unsigned short
+				result = epoch.AddDays ((ushort) comm.GetTdsShort ());
 				short minutes = comm.GetTdsShort ();
 				if (minutes != 0) 
 					result = ((DateTime) result).AddMinutes ((int) minutes);
